Renamed prismic-config to prismic-configuration
